Removed separate UUID generator class, and option for prefill.
