+++ src/Qowaiv/Uuid.cs
@@ -162,32 +162,37 @@ namespace Qowaiv
         /// * Withing a timespan of 32 ticks (0.32 nanoseconds) the sequential part
         /// of UUID's are identical.
         /// </remarks>
-        public static Uuid NewSequential(UuidComparer comparer) => NewSequential(comparer, false);
-
-        /// <summary>Initializes a new  instance of a UUID that is sequential.</summary>
-        /// <param name="comparer">
-        /// The comparer that determines the order to put the generated bytes in.
-        /// </param>
-        /// <param name="prefilled">
-        /// If the sequential UUID for a prefilled table/storage.
-        /// </param>
-        /// <remarks>
-        /// * The first 7 bytes are based on ticks from <see cref="Clock.UtcNow()"/>.
-        /// * The UUID's generated are sequential between 1970-01-01 and  9276-12-03.
-        /// * Withing a timespan of 32 ticks (0.32 nanoseconds) the sequential part
-        /// of UUID's are identical.
-        /// </remarks>
-        /// <remarks>
-        /// * The first 7 bytes are based on ticks from <see cref="Clock.UtcNow()"/>.
-        /// * The UUID's generated are sequential between 1970-01-01 and  9306-12-04.
-        /// * Withing a timespan of 32 ticks (0.32 nanoseconds) there is a 25% change
-        ///   that UUID's are not sequential.
-        /// * For preffilled UUID's the timespan is 8192 ticks (81.92 nanoseconds).
-        /// </remarks>
-        public static Uuid NewSequential(UuidComparer comparer, bool prefilled)
+        public static Uuid NewSequential(UuidComparer comparer)
         {
-            return UuidGenerator.Sequential(comparer, prefilled);
+            var prioritizer = (comparer ?? UuidComparer.Default).Priority;
+            var sequential = (Clock.UtcNow().Ticks - TicksYear1970) >> 5;
+
+            // replace the version byte with a fully random one.
+            var random = Guid.NewGuid().ToByteArray();
+            random[IndexOfVersion] = random[0];
+
+            // set the sequential part.
+            for (var i = 6; i >= 0; i--)
+            {
+                random[i] = (byte)sequential;
+                sequential >>= 8;
+            }
+
+            var bytes = new byte[ArraySize];
+
+            // setting the priority.
+            for (var index = 0; index < ArraySize; index++)
+            {
+                var prio = prioritizer[index];
+                bytes[prio] = random[index];
+            }
+
+            // setting the version.
+            UuidExtensions.SetVersion(bytes, UuidVersion.Sequential);
+
+            return new Guid(bytes);
         }
+        private const long TicksYear1970 = 0x89F_7FF5_F7B5_8000;
 
         /// <summary>Converts the string to a UUID.
         /// A return value indicates whether the conversion succeeded.
